Improve code organization: extract ConnectionMonitor, rename DeviceInfoUtils
