+++ library/src/main/java/com/wultra/android/powerauth/networking/utils/AppUtils.kt
@@ -19,10 +19,7 @@ package com.wultra.android.powerauth.networking.utils
 import android.content.Context
 import android.content.pm.PackageInfo
 import android.content.pm.PackageManager
-import android.net.*
 import android.os.Build
-import androidx.annotation.RequiresApi
-import com.wultra.android.powerauth.networking.log.WPNLogger
 
 class AppUtils {
     companion object {
@@ -39,57 +36,3 @@ class AppUtils {
     }
 }
 
-object ConnectionMonitor {
-
-    enum class Status(val value: String) {
-        UNKNOWN("unknown"),
-        NO_CONNECTION("noConnection"),
-        WIFI("wifi"),
-        CELLULAR("cellular"),
-        WIRED("wired")
-    }
-
-    fun getConnectivityStatus(context: Context): String {
-        var status: Status
-        try {
-            val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-            status = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                getConnectionStatus(connectivityManager)
-            } else {
-                val activeNetworkInfo = connectivityManager.activeNetworkInfo
-                if (activeNetworkInfo != null && activeNetworkInfo.isConnected) {
-                    getOlderConnectionStatus(activeNetworkInfo)
-                } else {
-                    Status.UNKNOWN
-                }
-            }
-        } catch (e: Throwable) {
-            status = Status.UNKNOWN
-            WPNLogger.d("Failed to create Connectivity Manager with Exception: $e")
-        }
-        return status.value
-    }
-
-    @RequiresApi(Build.VERSION_CODES.M)
-    private fun getConnectionStatus(connectivityManager: ConnectivityManager): Status {
-        val activeNetwork = connectivityManager.activeNetwork
-        val networkCapabilities = connectivityManager.getNetworkCapabilities(activeNetwork)
-
-        return when {
-            networkCapabilities == null -> Status.NO_CONNECTION
-            networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> Status.WIFI
-            networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> Status.CELLULAR
-            networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> Status.WIRED
-            else -> Status.UNKNOWN
-        }
-    }
-
-    private fun getOlderConnectionStatus(networkInfo: NetworkInfo): Status {
-        return when (networkInfo.type) {
-            ConnectivityManager.TYPE_WIFI -> Status.WIFI
-            ConnectivityManager.TYPE_MOBILE -> Status.CELLULAR
-            ConnectivityManager.TYPE_ETHERNET -> Status.WIRED
-            else -> Status.UNKNOWN
-        }
-    }
-}
