+++ library/src/main/java/com/wultra/android/powerauth/networking/utils/ConnectionMonitor.kt
@@ -1,0 +1,83 @@
+package com.wultra.android.powerauth.networking.utils
+
+/*
+ * Copyright 2026 Wultra s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.net.NetworkInfo
+import android.os.Build
+import androidx.annotation.RequiresApi
+import com.wultra.android.powerauth.networking.log.WPNLogger
+
+object ConnectionMonitor {
+
+    enum class Status(val value: String) {
+        UNKNOWN("unknown"),
+        NO_CONNECTION("noConnection"),
+        WIFI("wifi"),
+        CELLULAR("cellular"),
+        WIRED("wired")
+    }
+
+    fun getConnectivityStatus(context: Context): String {
+        var status: Status
+        try {
+            val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+            status = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                getConnectionStatus(connectivityManager)
+            } else {
+                @Suppress("DEPRECATION")
+                val activeNetworkInfo = connectivityManager.activeNetworkInfo
+                if (activeNetworkInfo != null && activeNetworkInfo.isConnected) {
+                    @Suppress("DEPRECATION")
+                    getOlderConnectionStatus(activeNetworkInfo)
+                } else {
+                    Status.UNKNOWN
+                }
+            }
+        } catch (e: Throwable) {
+            status = Status.UNKNOWN
+            WPNLogger.d("Failed to create Connectivity Manager with Exception: $e")
+        }
+        return status.value
+    }
+
+    @RequiresApi(Build.VERSION_CODES.M)
+    private fun getConnectionStatus(connectivityManager: ConnectivityManager): Status {
+        val activeNetwork = connectivityManager.activeNetwork
+        val networkCapabilities = connectivityManager.getNetworkCapabilities(activeNetwork)
+
+        return when {
+            networkCapabilities == null -> Status.NO_CONNECTION
+            networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> Status.WIFI
+            networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> Status.CELLULAR
+            networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> Status.WIRED
+            else -> Status.UNKNOWN
+        }
+    }
+
+    @Suppress("DEPRECATION")
+    private fun getOlderConnectionStatus(networkInfo: NetworkInfo): Status {
+        return when (networkInfo.type) {
+            ConnectivityManager.TYPE_WIFI -> Status.WIFI
+            ConnectivityManager.TYPE_MOBILE -> Status.CELLULAR
+            ConnectivityManager.TYPE_ETHERNET -> Status.WIRED
+            else -> Status.UNKNOWN
+        }
+    }
+}
